% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{check_trial_info}
\alias{check_trial_info}
\alias{read_apsimx_report}
\alias{get_hwsd2}
\alias{impute}
\alias{cv}
\alias{fct_contr_sum}
\alias{is_environmental_data}
\alias{create_MET}
\title{Hidden functions generally not to be directly called by the user}
\usage{
check_trial_info(trial.info)

read_apsimx_report(dir, apsim.name)

get_hwsd2(lat, long, gridsize, hwsd.raster, con)

impute(x)

cv(x, na.rm = FALSE)

fct_contr_sum(x, drop.levels = FALSE)

is_environmental_data(x)

create_MET(env.data, dir = ".")
}
\description{
Hidden functions generally not to be directly called by the user
}
\section{Functions}{
\itemize{
\item \code{read_apsimx_report}: 

\item \code{get_hwsd2}: 

\item \code{impute}: 

\item \code{cv}: 

\item \code{fct_contr_sum}: 

\item \code{is_environmental_data}: Checks if an object contains weather or soil data from the functions \code{\link{get_weather_data}},
\code{\link{get_historical_weather_data}}, or \code{\link{get_soil_data}}.

\item \code{create_MET}: 
}}

