% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{crossv_loo_grouped}
\alias{crossv_loo_grouped}
\alias{crossv_mc_grouped}
\alias{rapid_crossv}
\title{Ancillary cross-validation functions}
\usage{
crossv_loo_grouped(data, id = ".id")

crossv_mc_grouped(data, n, test = 0.25, id = ".id")

rapid_crossv(object, cv.indices, rapid)
}
\arguments{
\item{data}{A grouped data frame.}

\item{id}{Name of variable that gives each training set a unique integer id.}

\item{n}{The number of cross validation iterations.}

\item{test}{The proportion of data to save for model testing.}

\item{object}{A fitted model object}

\item{cv.indices}{A list of indices that define each training set. Metrics are calculated using the predicted and observed values from the holdout sets (i.e. opposite the training set indicies). Metrices will be averaged over multiple nested lists.}

\item{rapid}{Logical; should the rapid method be used?}
}
\description{
Ancillary cross-validation functions
}
\details{
Exported from the \href{https://github.com/neyhartj/neyhart}{neyhart} package.
}
\section{Functions}{
\itemize{
\item \code{crossv_mc_grouped}: 

\item \code{rapid_crossv}: 
}}

