% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_model.R
\name{define_growth_stages}
\alias{define_growth_stages}
\title{Define crop growth stages}
\usage{
define_growth_stages(env.data, growth.stage.delim, stages)
}
\arguments{
\item{env.data}{A \code{data.frame} of trials with daily environmental data, as created from \code{\link{get_weather_data}} or \code{\link{get_historical_weather_data}} (or another function that uses this data as an input, e.g. \code{\link{run_apsim_crop_model}}. The variable you want to use to delineate growth stages must be contained in the 'data' element of this object.}

\item{growth.stage.delim}{The variable you want to use to delineate growth stages in the 'data' element of \code{env.data}. For example, this may be the day of the year, or Zadok numeric stages, or growing degree days. It is up to the user to make sure that this variable is present.}

\item{stages}{A named list with \code{character} expressions defining the growth stages based off the \code{growth.stage.delim} variable.  See \emph{Details} for examples.}
}
\description{
Uses the output of a crop growth model, or some other user-defined quantity,
to assign crop growth stages to environmental data.
}
\details{
The elements of \code{stages} must be parsable logical expressions using the dummy variable "x," which will automatically be converted to the variable specified in \code{growth.stage.delim}. The rows in the 'data' elements of \code{env.data} that meet the criteria specified in \code{stages} will be designated to the appropriate growth stage, as given by the names of \code{stages}. For example, if \code{growth.stage.delim = "zadok_stage"} and \code{stages = list(stage1 = "x >= 1 & x <= 10", stage2 = "x > 10 & x <= 20")}, then rows in the 'data' element in which 'zadok_stage' is >= 1 or <= 10 will be designated as 'stage1', and rows in which 'zadok_stage' is > 10 and <= 20 will be designated as 'stage2' and so on.
}
\examples{
\dontrun{
# Get weather data for each trial
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")

# Run a crop model for each trial
# Use the zadok_stage to assign growth stages
trial_crop_growth <- run_apsim_crop_model(env.data = trial_weather,base.model.path = "barley")

barley_stages <- list(
  early_vegetative = "x >= 10 & x <= 30", late_vegetative = "x > 30 & x <= 50",
  flowering = "x > 50 & x <= 70", grain_fill = "x > 70 & x <= 91"
)

trial_crop_stages <- define_growth_stages(env.data = trial_crop_growth, growth.stage.delim = "zadok",
                                          stages = barley_stages)

}


}
