% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_environmental_data.R
\name{ec_matrix}
\alias{ec_matrix}
\title{Create an environmental covariable matrix}
\usage{
ec_matrix(
  env.data,
  env.col,
  var.col = "variable",
  val.col = "value",
  check.data = TRUE,
  alpha = 0.05,
  scale.data = TRUE
)
}
\arguments{
\item{env.data}{A \code{data.frame} of environmental variables in long format.}

\item{env.col}{The column in \code{env.data} that designates environments.}

\item{var.col}{The column in \code{env.data} that designates the names of the variables.}

\item{val.col}{The column in \code{env.data} that designates the value of the variables.}

\item{check.data}{Logical. Should a check for normality of each variables be conducted?}

\item{alpha}{The significance level at which to reject the null hypothesis of the normality test.}

\item{scale.data}{Logical. Should the data matrix be centered and scaled prior to output?}
}
\description{
Creates a \code{n x p} matrix for \code{n} environments and \code{p} covariables for use in modeling or to create environmental relationship matrices. Values of the covariables are centered and scaled.
}
\details{

}
\examples{
\dontrun{
# Get weather data for each trial
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")
# Get soil data for each trial
hwsd_path <- "path/to/hwsd.bil"
trial_soil <- get_soil_data(trial.info = trial_info, hwsd.path = hwsd_path)

# Run a crop model for each trial
trial_crop_growth <- run_apsim_crop_model(env.data = trial_weather, base.model.path = "barley")

# Use the zadok_stage to assign growth stages
barley_stages <- stages <- list(
  early_vegetative = "x >= 10 & x <= 30", late_vegetative = "x > 30 & x <= 50",
  flowering = "x > 50 & x <= 70", grain_fill = "x > 70 & x <= 91"
)

# Assign growth stages using the zadok stage
trial_crop_stages <- define_growth_stages(env.data = trial_crop_growth, growth.stage.delim = "zadok_stage",
                                          stages = barley_stages)

# A list of funs for summary
summary_funs <- list(rain = ~sum, radn = ~sum, maxt = ~mean, mint = ~mean, daylength = ~mean)

# Summarize environmental data according to the growth stages
trial_data_summarized <- summarize_environmental_data(
  env.data = trial_crop_stages,
  .funs = summary_funs,
  unite = TRUE
)

# Merge the summaried weather data and the soil data
env_data_merged <- merge_environmental_data(trial_data_summarized, trial_soil)

# Use tidyr to reshape the data
library(tidyr)
env_data_merged1 <- env_data_merged \%>\%
  unnest(data) \%>\%
  gather(variable, value, c("latitude", "longitude", "elevation",
                            names(env_data_merged$data[[1]])))

# Create a matrix of n environments by p covariables
E_matrix <- ec_matrix(env.data = env_data_merged1, env.col = "trial",
                      var.col = "variable", val.col = "value")

}

}
