% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromental_relationship.R
\name{ec_relmat}
\alias{ec_relmat}
\title{Construct an environmental "relationship" matrix from environmental covariables}
\usage{
ec_relmat(
  env.data,
  ec.matrix,
  ec.weights = NULL,
  method = c("method1", "method2", "method3"),
  ...
)
}
\arguments{
\item{env.data}{A \code{data.frame} of environmental variables in long format. Must be provided if \code{ec.matrix} is not passed; ignored otherwise. Must include information on the environments/trials/location in \code{pheno.data}.}

\item{ec.matrix}{A \code{matrix} of (ideally) centered and scaled environmental covariable data (i.e. from the \code{\link{ec_matrix}} function. Supercedes \code{env.data} if passed. Must include information on the environments/trials/location in \code{pheno.data}.}

\item{ec.weights}{A \code{matrix} of weights for individual covariables.}

\item{method}{The method of calculating the relationship matrix. See \emph{Details}.}

\item{...}{Other arguments to pass to the \code{\link{ec_matrix}} function (if \code{ec.matrix} is not passed).}
}
\description{
Construct an environmental "relationship" matrix from environmental covariables
}
\details{
The methods of calculaing the relationship matrix are defined as follows. In each, \emph{X} represents the centered and scaled \emph{n} x \emph{p} matrix of \emph{p} covariables for \emph{n} environments, and \eqn{Z = Xw}, where \emph{w} is a matrix of weights from \code{ec.weights}.

\itemize{
  \item{\code{method1}: \eqn{ (ZZ') / p } }
  \item{\code{method2}: \eqn{ 1 - \sum^p_{i=1}\frac{z_{ij} - z_{ij'}}{max(z_i) - min(z_i)} } }
  \item{\code{method3}: \eqn{ \frac{dist(Z)}{max(dist(Z))} } }

}
}
\examples{

env.data <- env_data_merged

# Example ec.weights from running ec_variable_selection
optVariables <- matrix(
  data = 1, nrow = 4,
  dimnames = list(c("flowering.daylength_mean", "grain_fill.radn_sum",
                    "subsoil_ref_bulk_density", "late_vegetative.radn_sum"),
                  "importance"))

# Calculate the relationship matrix from all covariables
Emat <- ec_relmat(env.data = env.data, env.col = "trial", var.col = "variable",
                  val.col = "value")

# Calculate the relationship matrix using a data.frame
# Use only the optimized variables from 'optVariables'
Emat <- ec_relmat(env.data = env.data, ec.weights = optVariables,
                  env.col = "trial", var.col = "variable", val.col = "value")

# Calculate a n x p matrix of covariables first, then use this to calculate
# the relationship matrix
ec_mat <- ec_matrix(env.data = env.data, env.col = "trial", var.col = "variable",
                    val.col = "value")
Emat <- ec_relmat(ec.matrix = ec_mat, ec.weights = optVariables)




}
\references{
Jarquin, D., J. Crossa, X. Lacaze, P. Du Cheyron, J. Daucourt, J. Lorgeou, F. Piraux, L. Guerreiro, P. Perez, M. Calus, J. Burgueno, and G. de los Campos. 2014. A reaction norm model for genomic selection using high-dimensional genomic and environmental data. Theor. Appl. Genet. 127(3): 595-607. doi: 10.1007/s00122-013-2243-1.

Malosetti, M., D. Bustos-Korts, M.P.P. Boer, and F. a. van Eeuwijk. 2016. Predicting Responses in Multiple Environments: Issues in Relation to Genotype x Environment Interactions. Crop Sci. 56: 2210-2222. doi: 10.2135/cropsci2015.05.0311.

Rincent, R., M. Malosetti, B. Ababaei, G. Touzy, A. Mini, M. Bogard, P. Martre, J. Le Gouis, and F. van Eeuwijk. 2019. Using crop growth model stress covariates and AMMI decomposition to better predict genotype-by-environment interactions. Theor. Appl. Genet. 132(12): 3399-3411. doi: 10.1007/s00122-019-03432-y.
}
