% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_environmental_data.R
\name{ec_variable_selection}
\alias{ec_variable_selection}
\alias{ec_variable_selection_slide}
\title{Select environmental covariables based on phenotype data}
\usage{
ec_variable_selection(
  pheno.data,
  env.data,
  ec.matrix,
  env.col,
  gen.col,
  y.col,
  ec.subset = NULL,
  search.method = c("stepwise.lmm", "lasso"),
  cv.type = c("loo", "mc"),
  cv.ptrain = 0.6,
  cv.iter = 25,
  ...
)

ec_variable_selection_slide(
  pheno.data,
  env.data,
  env.col,
  gen.col,
  y.col,
  var.col,
  val.col,
  ec.subset = NULL,
  search.method = c("stepwise.lmm", "lasso"),
  time.col,
  time.list,
  cv.type = c("loo", "mc"),
  cv.ptrain = 0.6,
  cv.iter = 25
)
}
\arguments{
\item{pheno.data}{A \code{data.frame} of phenotype data that will be used in the analysis or to train predictive models. This must contain columns for genotypes, environments, and the value of the phenotypic observation.}

\item{env.data}{A \code{data.frame} of environmental variables in long format. Must be provided if \code{ec.matrix} is not passed; ignored otherwise. Must include information on the environments/trials/location in \code{pheno.data}.}

\item{ec.matrix}{A \code{matrix} of (ideally) centered and scaled environmental covariable data (i.e. from the \code{\link{ec_matrix}} function. Supercedes \code{env.data} if passed. Must include information on the environments/trials/location in \code{pheno.data}.}

\item{env.col}{The column in \code{pheno.data} and (if passed) \code{env.data} that designates environments.}

\item{gen.col}{The column in \code{pheno.data} that designates genotypes.}

\item{y.col}{The column in \code{pheno.data} that designates phenotypic values.}

\item{ec.subset}{A \code{character} vector of covariables in \code{env.data} or \code{ec.matrix} that should be considered. If \code{NULL} (default), all covariables in \code{env.data} or \code{ec.matrix} are considered.}

\item{search.method}{The covariable selection search method. See \emph{Details} for more information.}

\item{cv.type}{The type of cross-validation used to select variables. Can be \code{"loo"} for leave-one-out (environments) or \code{"mc"} for random train-test partitions.}

\item{cv.ptrain}{The proportion of data to use in model training if \code{cv.type = "mc"}.}

\item{cv.iter}{The number of cross-validation replications if \code{cv.type = "mc"}.}

\item{...}{Other arguments to pass to the \code{\link{ec_matrix}} function (if \code{ec.matrix} is not passed).}

\item{var.col}{See \code{\link{ec_matrix}}.}

\item{val.col}{See \code{\link{ec_matrix}}.}

\item{time.col}{The column in \code{env.data} used to create time frames (this is usually year).}

\item{time.list}{A list of vectors, each containing elements of \code{time.col} that delineate the timeframes. If not passed, a list is created that accumulates elements of \code{time.col} from maximum to minimum in steps of 1 unit.}
}
\value{
The function \code{ec_variable_selection} returns a list with the following elements:

\describe{
  \item{\code{optVariables}}{A n x 1 \code{matrix} where the row names are the n environmental covariables selected and the elements of the matrix are the importance value of that variable.}
  \item{\code{finalCVResults}}{A vector with the cross-validation results of the final model.}
  \item{\code{CVStepResults}}{A matrix detailing the cross-validation results at each step of the stepwise linear model variable selection approach (i.e. \code{search.method = "stepwise.lmm"})}

}

The function \code{ec_variable_selection_slide} returns a list with the following elements:

\describe{
  \item{\code{timeframeFinalCVResults}}{A \code{data.frame} of cross-validation results for each timeframe.}
  \item{\code{variableSelectionResults}}{A list containing the same elements as the output of \code{ec_variable_selection}, but only for the optimal timeframe.}
  \item{\code{ecMatrixSelected}}{The n x p matrix of environmental covariables summarized across the optimal timeframe.}

}
}
\description{
Perform variable selection using observed phenotype data to determine the most relevant set of environmental covariables
}
\details{
TBD.
}
\examples{
\dontrun{

pheno.data <- pheno_data_train
env.data <- env_data_merged

variable_selection_out <- ec_variable_selection(
  pheno.data = pheno.data, env.data = env.data, env.col = "trial",
  gen.col = "line_name", y.col = "value", var.col = "variable",
  val.col = "value"
)

}


\dontrun{

# Replace location
pheno.data <- merge(pheno_data_train[-2], trial_info[c("trial", "location")])
env.data <- hist_env_data_merged

variable_selection_out <- ec_variable_selection_slide(
  pheno.data = pheno.data, env.data = env.data, env.col = "location",
  gen.col = "line_name", y.col = "value", var.col = "variable",
  val.col = "value", time.col = "year"
)

}


}
