% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_environmental_data.R
\name{edit_environmental_data}
\alias{edit_environmental_data}
\title{Edit environmental data}
\usage{
edit_environmental_data(env.data, suppl.data, operation = c(all = "+"))
}
\arguments{
\item{env.data}{A \code{data.frame} of environmental data from the outputs of \code{\link{get_weather_data}},
\code{\link{get_historical_weather_data}}, or \code{\link{get_soil_data}}.}

\item{suppl.data}{A \code{data.frame} with data to ammend to the \code{env.data} input.}

\item{operation}{A named vector specifying the editing operation for each variable. For instance, if
\code{operation = c("rain" = "add")}, rainfall information in \code{env.data} will be edited by
adding the rainfall data from \code{suppl.data}. Operations can include any accepted mathematical operation (e.g. "+", "-", "*")
or "replace" indicating that values will be edited by replacement.}
}
\description{
Allows simple edits to the environmental data based on a complementary data.frame. This could be useful for, say,
ammending rainfal information with irrigation data.
}
\examples{
\dontrun{
# Get daily weather data
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")

# Modify the trial irrigation data to set irrigated water as "rain"
trial_irrigation_data$rain <- trial_irrigation_data$irrigation_mm

# Edit the trial weather data with this irrigation information
trial_weather1 <- edit_environmental_data(env.data = trial_weather, suppl.data = trial_irrigation_data)
}


}
