% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotyping_package.R
\docType{package}
\name{envirotyping}
\alias{envirotyping}
\title{envirotyping: A package for obtaining useful environmental covariables for modeling in quantitative genetics and genomics}
\section{Functions}{

The foo functions ...
}

\section{Crop Model Examples}{

The package is distributed with some example base APSIM crop model files. These include ...
}

\author{
Jeff Neyhart (creator)
}
