% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_environmental_data.R
\name{get_historical_weather_data}
\alias{get_historical_weather_data}
\title{Query public weather data over longer timeframes}
\usage{
get_historical_weather_data(
  location.info,
  start.year,
  end.year,
  source = c("daymet", "nasapower"),
  verbose = TRUE
)
}
\arguments{
\item{location.info}{A list of location metadata that includes the location name,
latitude, and longitude. This can be a \code{data.frame}. See \emph{Details} for specific information.}

\item{start.year}{The first year to gather data. (Note that the maximum year range between \code{start.year} and \code{end.year} is 30.)}

\item{end.year}{The last year to gather data. (Note that the maximum year range between \code{start.year} and \code{end.year} is 30.)}

\item{source}{The database to query. Can be one of \code{"nasapower"} or \code{"daymet"}.
See \emph{Details} for more information.}

\item{verbose}{Logical. Should progress messages be printed?}
}
\description{
Queries long-term weather data from publicly available datasets based on geographical coordinates.
}
\details{

}
\examples{
\dontrun{
# Find distinct location info
location_info <- dplyr::distinct(trial_info, location, latitude, longitude)

# Retrieve historical weather data from 2005-2014 (10 years)
location_historical_weather <- get_historical_weather_data(
  location.info = location_info, start.year = 2005,
  end.year = 2014, source = "daymet"
)
}


}
