% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_environmental_data.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Query public weather data}
\usage{
get_weather_data(trial.info, source = c("daymet", "nasapower"), verbose = TRUE)
}
\arguments{
\item{trial.info}{A list of trial metadata that includes the trial name, year,
latitude, and longitude. This can be a \code{data.frame}. See \emph{Details} for specific information.}

\item{source}{The database to query. Can be one of \code{"nasapower"} or \code{"daymet"}.
See \emph{Details} for more information.}

\item{verbose}{Logical. Should progress messages be printed?}
}
\description{
Queries weather data from publicly available datasets based on geographical coordinates.
}
\examples{
\dontrun{
# Use the included trial_info data
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")
}


}
