% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotyping_package_data.R
\docType{data}
\name{hwsd.variables}
\alias{hwsd.variables}
\title{Variable names from the Harmonic World Soil Database (HWSD)}
\format{
A \code{data.frame} with 60 rows and 5 variables:

\describe{
  \item{field}{The variable name as expressed in the HWSD}
  \item{type}{The type of variable, one of "general", "other", or "property"}
  \item{variable}{A variable nickname}
  \item{units}{The variable units}
  \item{unitsR}{The variable units formatted for use in an expression}
}
}
\usage{
hwsd.variables
}
\description{
Variable names from the Harmonic World Soil Database (HWSD)
}
\keyword{datasets}
