% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{lasso_fi}
\alias{lasso_fi}
\title{Determine feature importance using the LASSO procedure}
\usage{
lasso_fi(object, data, env.col, scope, metric = c("RMSE", "R2"), cv.indices)
}
\description{
Determine feature importance using the LASSO procedure
}
