% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_environmental_data.R
\name{merge_environmental_data}
\alias{merge_environmental_data}
\title{Merge environmental data}
\usage{
merge_environmental_data(...)
}
\arguments{
\item{...}{Two or more \code{data.frames} of environmental data. Note that elements of the 'data' column in each data frame must share either a row or column dimension.}
}
\description{
Merge two or more sources of environmental data to create a single \code{data.frame}.
}
\examples{
\dontrun{
# Get weather data for each trial
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")
# Get soil data for each trial
hwsd_path <- "path/to/hwsd.bil"
trial_soil <- get_soil_data(trial.info = trial_info, hwsd.path = hwsd_path)

# Run a crop model for each trial
trial_crop_growth <- run_apsim_crop_model(env.data = trial_weather, base.model.path = "barley")

# Use the zadok_stage to assign growth stages
barley_stages <- stages <- list(
  early_vegetative = "x >= 10 & x <= 30", late_vegetative = "x > 30 & x <= 50",
  flowering = "x > 50 & x <= 70", grain_fill = "x > 70 & x <= 91"
)

# Assign growth stages using the zadok stage
trial_crop_stages <- define_growth_stages(env.data = trial_crop_growth, growth.stage.delim = "zadok_stage",
                                          stages = barley_stages)

# A list of funs for summary
summary_funs <- list(rain = ~sum, radn = ~sum, maxt = ~mean, mint = ~mean, daylength = ~mean)

# Summarize environmental data according to the growth stages
trial_data_summarized <- summarize_environmental_data(
  env.data = trial_crop_stages,
  .funs = summary_funs,
  unite = TRUE
)

# Merge the summaried weather data and the soil data
env_data_merged <- merge_environmental_data(trial_data_summarized, trial_soil)

}

}
