% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_model.R
\name{run_apsim_crop_model}
\alias{run_apsim_crop_model}
\title{Run APSIM crop model using current or historical weather data}
\usage{
run_apsim_crop_model(
  env.data,
  wd = ".",
  base.model.path,
  exe.path,
  verbose = TRUE
)
}
\arguments{
\item{env.data}{A \code{data.frame} of environmental data from the outputs of \code{\link{get_weather_data}}
or \code{\link{get_historical_weather_data}}. This must include planting date information.}

\item{wd}{The working directory. This function will write and read files, so it is recommended that you set an appropriate working directory.}

\item{base.model.path}{The path to a base crop model file (\code{.apsimx}) or the name of a crop model example
that is distributed in this package.}

\item{exe.path}{The path to the APSIM executable. If missing, the function will try to
guess where the executable is.}

\item{verbose}{Logical. Should progress messages be printed?}
}
\description{
Uses the APSIM software to run crop models based on user-supplied environmental data.
}
\details{
This function relies on the user having an installation of the \href{https://www.apsim.info/apsim-next-generation/}{APSIM Next Generation} software, which is freely available for academic use. Make sure you install this software prior to using this function.


This function relies heavily on the \code{\href{https://cran.r-project.org/package=apsimx}{apsimx}} package.
}
\examples{
\dontrun{
# Get weather data for each trial
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")

# Run a crop model for each trial
trial_crop_growth <- run_apsim_crop_model(env.data = trial_weather, base.model.path = "barley")

}


}
