% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_environmental_data.R
\name{save_ec_selection_table}
\alias{save_ec_selection_table}
\title{Output a table of selected environmental covariates}
\usage{
save_ec_selection_table(x, file = NULL)
}
\arguments{
\item{x}{The output object of the \code{\link{ec_variable_selection}} or \code{\link{ec_variable_selection_slide}} functions.}

\item{file}{File or connection to write to. If NULL (default), the function simply returns a tidy data frame.}
}
\description{
Saves a table of the output from the \code{\link{ec_variable_selection}} or \code{\link{ec_variable_selection_slide}} functions.
}
