% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{stepwise_rfa}
\alias{stepwise_rfa}
\title{Functions for environmental covariable selection}
\usage{
stepwise_rfa(
  object,
  data,
  env.col,
  scope,
  metric = c("RMSE", "R2"),
  cv.indices
)
}
\arguments{
\item{object}{The base fitted model object. Can be a \code{lm} or \code{merMod} object.}

\item{data}{The data frame that includes phenotypic observations and values for the environmental covariables.}

\item{env.col}{The column in \code{data} corresponding to environments.}

\item{scope}{A named list of two formulas defining the lower and upper scope of the variable search.}

\item{metric}{The metric to optimize: either minimize the root mean squared error (\code{"RMSE"}) or maximize the coefficient of determination (\code{"R2"}).}

\item{cv.indices}{A list of indices that define each training set. Metrics are calculated using the predicted and observed values from the holdout sets (i.e. opposite the training set indicies). Metrices will be averaged over multiple nested lists.}
}
\description{
These functions provide different approaches to selecting environmental covariables for use in quantitative genetic analysis or prediction. \code{stepwise_rfa} implements a stepwise recursive feature addition algorithm based on optimizing a metric (the root mean squared error or the coefficient of determination) by sequentially adding (or removing) variables.
}
\details{
Note that inputs for this function are only sparsely checked for correctness.
}
