% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_environmental_data.R
\name{summarize_environmental_data}
\alias{summarize_environmental_data}
\title{Summarize environmental data according to growth stages}
\usage{
summarize_environmental_data(
  env.data,
  growth.stage.col = "growth_stage",
  .funs,
  ...,
  unite = FALSE
)
}
\arguments{
\item{env.data}{A \code{data.frame} of trials with daily environmental data and categorical growth stage information, as created by \code{\link{define_growth_stages}}.}

\item{growth.stage.col}{The column in the elements of 'data' in \code{env.data} that contains the categorical growth stage information.}

\item{.funs}{A named list of formulas specifying the summary function for a specific variable (the names of \code{.funs}). Only the variables listed in the names of \code{.funs} are summarized.}

\item{...}{Other arguments to pass to the functions listed in \code{.funs} (e.g. \code{na.rm = TRUE}).}

\item{unite}{Logical. Should growth stages and variable names be combined?}
}
\description{
Use categorical growth stage data to summarize quantitative environmental variables.
}
\examples{
\dontrun{
# Get weather data for each trial
trial_weather <- get_weather_data(trial.info = trial_info, source = "daymet")

# Run a crop model for each trial
trial_crop_growth <- run_apsim_crop_model(env.data = trial_weather, base.model.path = "barley")

# Use the zadok_stage to assign growth stages
barley_stages <- stages <- list(
  early_vegetative = "x >= 10 & x <= 30", late_vegetative = "x > 30 & x <= 50",
  flowering = "x > 50 & x <= 70", grain_fill = "x > 70 & x <= 91"
)

# Assign growth stages using the zadok stage
trial_crop_stages <- define_growth_stages(env.data = trial_crop_growth, growth.stage.delim = "zadok_stage",
                                          stages = barley_stages)

# A list of funs for summary
summary_funs <- list(rain = ~sum, radn = ~sum, maxt = ~mean, mint = ~mean, daylength = ~mean)

# Summarize environmental data according to the growth stages
trial_data_summarized <- summarize_environmental_data(env.data = trial_crop_stages, .funs = summary_funs)

}

}
