% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_prediction.R
\name{tidy.mmer}
\alias{tidy.mmer}
\title{Functions to enable genomic prediction using environmental data}
\usage{
tidy.mmer(object)
}
\description{
Functions to enable genomic prediction using environmental data
}
